USB functionality
it actually connects to computers and identifies as a HID and translates its stored buffer to what a generic USB HID gamepad should be outputting, regardless of what the actual hardware configuration is. it really wants to be an xbox controller and translates whatever modules are plugged in to fit that. 

todo: saving custom configs, bluetooth connectivity, mapping the various buttons that are not the dpad or whatever for the generic USB gamepad object